+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v5.3.0
 ## 🚀 Feature
 - Added internal method to animate `AndesButton` transitions from different hierarchies | Author: [@juanunzu](https://github.com/juanunzu)
+- Add label formatter for AndesSlider
 
 # v5.2.0
 ## 🚀 Feature
@@ -14,6 +15,7 @@
 ## Migrated
 - Migration Kotlin 1.5.32 and Gradle 7.4. | Author: [@meligeo](https://github.com/meligeo)
 - Migration to the min API 23. | Author: [@meligeo](https://github.com/meligeo)
+
 ## 🚀 Feature
 - Added a switch to choose which day of the week it should start with on Andes DatePicker component
 - Added Iconography entry to docs | Author: [@eportasML](https://github.com/eportasML)